+++ tests/Method/SendDiceMethodTest.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+namespace TgBotApi\BotApiBase\Tests\Method;
+
+use TgBotApi\BotApiBase\BotApiComplete;
+use TgBotApi\BotApiBase\Method\SendContactMethod;
+use TgBotApi\BotApiBase\Method\SendDiceMethod;
+use TgBotApi\BotApiBase\Tests\Method\Traits\InlineKeyboardMarkupTrait;
+
+class SendDiceMethodTest extends MethodTestCase
+{
+    use InlineKeyboardMarkupTrait;
+
+    /**
+     * @throws \TgBotApi\BotApiBase\Exception\BadArgumentException
+     * @throws \TgBotApi\BotApiBase\Exception\ResponseException
+     */
+    public function testEncode()
+    {
+        $this->getApi()->sendDice($this->getMethod());
+        $this->getApi()->send($this->getMethod());
+    }
+
+    private function getApi(): BotApiComplete
+    {
+        return $this->getBot('sendDice', [
+            'chat_id' => 'chat_id',
+            'disable_notification' => true,
+            'reply_to_message_id' => 1,
+            'reply_markup' => $this->buildInlineMarkupArray(),
+        ], [], ['reply_markup']);
+    }
+
+    /**
+     * @throws \TgBotApi\BotApiBase\Exception\BadArgumentException
+     *
+     * @return SendContactMethod
+     */
+    private function getMethod(): SendDiceMethod
+    {
+        return SendDiceMethod::create(
+            'chat_id',
+            [
+                'disableNotification' => true,
+                'replyToMessageId' => 1,
+                'replyMarkup' => $this->buildInlineMarkupObject(),
+            ]
+        );
+    }
+}